==> 0_create.Rmd <==
---
output:
  html_notebook:
    code_folding: show
    toc: yes
    toc_float: yes
  html_document:
    df_print: paged
    toc: yes
editor_options: 
  chunk_output_type: inline
---

# Getting Data


***

## 1. Load... {.tabset}
### ...Packages
```{r message=FALSE, warning=FALSE}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(
  haven,
  lme4, 
  sjPlot,
  sjmisc,
  sjlabelled,
  ggthemes, 
  reshape, # reshape dataframes
  tmap, # to map spacial data, get shapefiles
  tmaptools, # to map spacial data
  ggmap, # to map spacial data
  sp, # to handle spacial data
  broom, # to clean up all kinds of datasets
  ggplot2, # to plot 
  gridExtra, # to arrange multiple plots
  tidyverse) # data wrangling

# wd <- "D:/Dropbox/DataVisualization" # Flo
# setwd(wd)
```

### ...Bitcoin Prices
```{r bitcoin prices, echo=T, message=F, warning=FALSE}
# Price of Bitcoins
      # price  <- read_csv("data/crypto-markets.csv")
      # saveRDS(price, "data/price.rds")
price <- readRDS("data/price.rds")
      # Relative Amount of Google Web Searches
      # google <- read_csv("data/Bitcoin_google_1_10_2017.csv")
# saveRDS(google, "data/google.rds")
google <- readRDS("data/google.rds")

# Relative Amount of News
      # news   <- read_csv("data/Bitcoin_news_1_10_2017.csv")
      # saveRDS(news, "data/news.rds")
news <- readRDS("data/news.rds")
```

### ...Bitcoin Region
[Info on Source of Data](https://support.google.com/trends/answer/4355212)
```{r bitcoin region, echo=T, message=F, warning=FALSE}
# Relative Amount of Google Web Searches by Region
## excludes low volume countries
gg_oct17_world <- read_csv("data/bitcoin_gg_region_011017.csv", skip = 3,
                           col_names = c("Country", "Bitcoin"))

## includes low volume countries
gg_oct17_world2 <- read_csv("data/bitcoin_gg_region_all_011017.csv", skip = 3,
                           col_names = c("Country", "Bitcoin"))

## shapefile of world
data(World)
```

## 2. Prepare Data {.tabset}
### Bitcoin Prices
#### Prepare Price Data
```{r}
price %>%
  filter(symbol == "BTC") %>%
  select(date, open, spread, market) ->
price
```

#### Merge 3 datasets
```{r}
price %>%
  right_join(google, by = "date") %>% 
  right_join(news, by = "date") ->
merge
```

#### Rename variables
```{r}
names(merge)[2:ncol(merge)] <- c("Preis", "Spread", "Gesamtwert", 
                                 "Google-Suchen", "Nachrichtenartikel")
```

#### Save Dataset
```{r}
saveRDS(merge, "data/bitcoin_raw.rds")
```


### Bitcoin Region

#### Merge Shapefile with Gogle Trends Regional Data
```{r}
world.shp <- sp::merge(World, gg_oct17_world, 
                       by.x = "name", by.y = "Country", all.x = T)
```
